// backend/src/modules/exercises/enums/exercise-equipment-type.enum.ts
export enum ExerciseEquipmentType {
  BODYWEIGHT = 'BODYWEIGHT',
  BARBELL = 'BARBELL',
  DUMBBELL = 'DUMBBELL',
  MACHINE = 'MACHINE',
  CABLE = 'CABLE',
  KETTLEBELL = 'KETTLEBELL',
  RESISTANCE_BAND = 'RESISTANCE_BAND',
  EZ_BAR = 'EZ_BAR',
  TRAP_BAR = 'TRAP_BAR',
  CARDIO_MACHINE = 'CARDIO_MACHINE',
  OTHER = 'OTHER',
}